Update README.md
The purpose of the PyBer_Analysis is to calculate the total weekly fares of the three city types, suburban, urban, and rural, using multiple-line graphs to demonstrate the outcomes. To achieve the results, we are asked to analyze the total rides of each city type, total drivers, the average fare per ride, and average fare per driver and compare the average fare to the total number of rides.

     

Total Rides	Total Drivers	Total Fares	Average Fare per Ride	Average Fare per Driver
type					
Rural	125	78	4327.93	34.623440	55.486282
Suburban	625	490	19356.33	30.970128	39.502714
Urban	1625	2405	39854.38	24.525772	16.571468
In [11]:



We are provided with two data, the city_data.csv and the ride_data.csv, to calculate the urban, suburban, and rural area total ride, total drivers, total fares, averages fare per ride, and the average fare per driver to graph a multiple-line chart graph. The graph shows that the total fare of the city from January to May. We realize that there is an increase in the total fare by the City Type at the end of February, the beginning of March, and in April. The output seems to level up from January to the beginning of April with a little bump up at the end of February. There is an increase during April, which may be associated with the summer season. There is also an increase in the number of rides and the total number of drivers in urban and suburban cities. It may be attributed to the number of people living in those cities. Usually, the number of people living in the Suburban and Urban cities is higher than the number of people living in the rural areas, which may lead to higher demand for rides, higher drivers and high prices, especially at a particular time of the day.

 



Summary: Overall, the Pyber summary and the multiple-line chart show an increase in total fares in April in Suburban cities, a decrease in rural and urban cities. The demand for the ride in populated cities than in the less populated may be attributed to reasons such as time of the years or the availability of other forms of transportation, or the number of people living in the cities. The information would help the company understand the fluctuation of the demand and supply of the driver needed in a particular time of the years. It may also help them increase the number of drivers required in a certain period of the years when they realize that the demand for their services is typically high.  Finally, the data is essential for the company to make a strategic decision to stay in business and make a profit in the long term. The data would also help the company to hire more drivers and understand the reason behind low rides in a specific area at a particular time of the year.